refactor: :recycle: 프로젝트 리스트 브랜치 컴포넌트 재활용할 여지 있으므로 분리
